fix(alembic): squash all migratons into one
