--- alembic/versions/3765e70eff36_migrate_to_update_answers_model.py
@@ -1,54 +1,0 @@
-"""migrate to update Answers model
-
-Revision ID: 3765e70eff36
-Revises: db3ef5a68698
-Create Date: 2026-05-20 07:05:49.660661
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '3765e70eff36'
-down_revision: Union[str, Sequence[str], None] = 'db3ef5a68698'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-	"""Upgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.execute(
-		"""
-        UPDATE answers
-        SET is_correct = 0
-        WHERE is_correct IS NULL
-        """
-	)
-	op.alter_column(
-		'answers',
-		'is_correct',
-		existing_type=sa.INTEGER(),
-		postgresql_using='is_correct::boolean',
-		type_=sa.Boolean(),
-		nullable=False,
-	)
-	# ### end Alembic commands ###
-
-
-def downgrade() -> None:
-	"""Downgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.alter_column(
-		'answers',
-		'is_correct',
-		existing_type=sa.Boolean(),
-		postgresql_using='is_correct::integer',
-		type_=sa.INTEGER(),
-		nullable=True,
-	)
-	# ### end Alembic commands ###
