--- alembic/versions/5f719406d96a_add_state_to_game_model.py
@@ -1,44 +1,0 @@
-"""add state to game model
-
-Revision ID: 5f719406d96a
-Revises: 76539444fff6
-Create Date: 2026-06-14 08:26:06.272579
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '5f719406d96a'
-down_revision: Union[str, Sequence[str], None] = '76539444fff6'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-	"""Upgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.add_column(
-		'games',
-		sa.Column(
-			'state',
-			sa.Enum(
-				'WAITING', 'ACTIVE', 'FINISHED', name='gamestate', native_enum=False, length=50
-			),
-			nullable=False,
-			server_default='WAITING',
-		),
-	)
-	op.alter_column('games', 'state', server_default=None)
-	# ### end Alembic commands ###
-
-
-def downgrade() -> None:
-	"""Downgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.drop_column('games', 'state')
-	# ### end Alembic commands ###
