--- alembic/versions/6757ec83b0c6_update_models_based_on_standards.py
@@ -1,173 +1,0 @@
-"""update models based on standards
-
-Revision ID: 6757ec83b0c6
-Revises: 5f719406d96a
-Create Date: 2026-06-17 05:32:30.664646
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = '6757ec83b0c6'
-down_revision: Union[str, Sequence[str], None] = '5f719406d96a'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-	"""Upgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.alter_column(
-		'answers',
-		'created_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=False,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'answers',
-		'updated_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=False,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'categories',
-		'created_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=False,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'categories',
-		'updated_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=False,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'games',
-		'created_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=False,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'games',
-		'updated_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=False,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'questions',
-		'created_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=False,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'questions',
-		'updated_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=False,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column('users', 'is_active', existing_type=sa.BOOLEAN(), nullable=False)
-	op.alter_column(
-		'users',
-		'created_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=False,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'users',
-		'updated_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=False,
-		existing_server_default=sa.text('now()'),
-	)
-	# ### end Alembic commands ###
-
-
-def downgrade() -> None:
-	"""Downgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.alter_column(
-		'users',
-		'updated_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=True,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'users',
-		'created_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=True,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column('users', 'is_active', existing_type=sa.BOOLEAN(), nullable=True)
-	op.alter_column(
-		'questions',
-		'updated_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=True,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'questions',
-		'created_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=True,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'games',
-		'updated_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=True,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'games',
-		'created_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=True,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'categories',
-		'updated_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=True,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'categories',
-		'created_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=True,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'answers',
-		'updated_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=True,
-		existing_server_default=sa.text('now()'),
-	)
-	op.alter_column(
-		'answers',
-		'created_at',
-		existing_type=postgresql.TIMESTAMP(timezone=True),
-		nullable=True,
-		existing_server_default=sa.text('now()'),
-	)
-	# ### end Alembic commands ###
