--- alembic/versions/76539444fff6_migrate_to_update_answers_model.py
@@ -1,110 +1,0 @@
-"""migrate to update Answers model
-
-Revision ID: 76539444fff6
-Revises: 3765e70eff36
-Create Date: 2026-06-13 10:43:05.845671
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '76539444fff6'
-down_revision: Union[str, Sequence[str], None] = '3765e70eff36'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-	"""Upgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.create_table(
-		'games',
-		sa.Column('id', sa.Integer(), nullable=False),
-		sa.Column('name', sa.String(length=100), nullable=False),
-		sa.Column('min_players', sa.Integer(), nullable=False),
-		sa.Column('max_players', sa.Integer(), nullable=False),
-		sa.Column('max_timeout', sa.Integer(), nullable=False),
-		sa.Column('category_id', sa.Integer(), nullable=False),
-		sa.Column(
-			'created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True
-		),
-		sa.Column(
-			'updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True
-		),
-		sa.Column('started_at', sa.DateTime(timezone=True), nullable=True),
-		sa.Column('ended_at', sa.DateTime(timezone=True), nullable=True),
-		sa.Column('creator_id', sa.Integer(), nullable=False),
-		sa.Column('private', sa.Boolean(), nullable=False),
-		sa.Column('password', sa.String(length=100), nullable=True),
-		sa.Column('winner', sa.Integer(), nullable=True),
-		sa.ForeignKeyConstraint(
-			['category_id'],
-			['categories.id'],
-		),
-		sa.ForeignKeyConstraint(
-			['creator_id'],
-			['users.id'],
-		),
-		sa.ForeignKeyConstraint(
-			['winner'],
-			['users.id'],
-		),
-		sa.PrimaryKeyConstraint('id'),
-	)
-	op.create_index(op.f('ix_games_id'), 'games', ['id'], unique=False)
-	op.create_table(
-		'game_players',
-		sa.Column('game_id', sa.Integer(), nullable=False),
-		sa.Column('user_id', sa.Integer(), nullable=False),
-		sa.ForeignKeyConstraint(['game_id'], ['games.id'], ondelete='CASCADE'),
-		sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-		sa.PrimaryKeyConstraint('game_id', 'user_id'),
-	)
-	op.create_table(
-		'game_questions',
-		sa.Column('game_id', sa.Integer(), nullable=False),
-		sa.Column('question_id', sa.Integer(), nullable=False),
-		sa.ForeignKeyConstraint(['game_id'], ['games.id'], ondelete='CASCADE'),
-		sa.ForeignKeyConstraint(['question_id'], ['questions.id'], ondelete='CASCADE'),
-		sa.PrimaryKeyConstraint('game_id', 'question_id'),
-	)
-	op.drop_constraint(op.f('answers_question_id_fkey'), 'answers', type_='foreignkey')
-	op.create_foreign_key(None, 'answers', 'questions', ['question_id'], ['id'], ondelete='CASCADE')
-	op.drop_constraint(op.f('questions_category_id_fkey'), 'questions', type_='foreignkey')
-	op.create_foreign_key(
-		None, 'questions', 'categories', ['category_id'], ['id'], ondelete='CASCADE'
-	)
-	# ### end Alembic commands ###
-
-
-def downgrade() -> None:
-	"""Downgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.drop_constraint(None, 'questions', type_='foreignkey')
-	op.create_foreign_key(
-		op.f('questions_category_id_fkey'),
-		'questions',
-		'categories',
-		['category_id'],
-		['id'],
-		ondelete='CASCADE',
-	)
-	op.drop_constraint(None, 'answers', type_='foreignkey')
-	op.create_foreign_key(
-		op.f('answers_question_id_fkey'),
-		'answers',
-		'questions',
-		['question_id'],
-		['id'],
-		ondelete='CASCADE',
-	)
-	op.drop_table('game_questions')
-	op.drop_table('game_players')
-	op.drop_index(op.f('ix_games_id'), table_name='games')
-	op.drop_table('games')
-	# ### end Alembic commands ###
