--- alembic/versions/bbbcf69ab6be_initial_migration.py
@@ -1,56 +1,0 @@
-"""Initial migration
-
-Revision ID: bbbcf69ab6be
-Revises:
-Create Date: 2026-05-18 06:35:26.383484
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'bbbcf69ab6be'
-down_revision: Union[str, Sequence[str], None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-	"""Upgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.create_table(
-		'users',
-		sa.Column('id', sa.Integer(), nullable=False),
-		sa.Column('email', sa.String(length=255), nullable=False),
-		sa.Column('phone_number', sa.String(length=16), nullable=True),
-		sa.Column('username', sa.String(length=100), nullable=False),
-		sa.Column('full_name', sa.String(length=255), nullable=True),
-		sa.Column('is_active', sa.Boolean(), nullable=True),
-		sa.Column(
-			'created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True
-		),
-		sa.Column(
-			'updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True
-		),
-		sa.PrimaryKeyConstraint('id'),
-	)
-	op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
-	op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-	op.create_index(op.f('ix_users_phone_number'), 'users', ['phone_number'], unique=True)
-	op.create_index(op.f('ix_users_username'), 'users', ['username'], unique=True)
-	# ### end Alembic commands ###
-
-
-def downgrade() -> None:
-	"""Downgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.drop_index(op.f('ix_users_username'), table_name='users')
-	op.drop_index(op.f('ix_users_phone_number'), table_name='users')
-	op.drop_index(op.f('ix_users_id'), table_name='users')
-	op.drop_index(op.f('ix_users_email'), table_name='users')
-	op.drop_table('users')
-	# ### end Alembic commands ###
