--- alembic/versions/db3ef5a68698_migrate_question_related_models.py
@@ -1,83 +1,0 @@
-"""migrate question related models
-
-Revision ID: db3ef5a68698
-Revises: bbbcf69ab6be
-Create Date: 2026-05-18 08:14:50.694121
-
-"""
-
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'db3ef5a68698'
-down_revision: Union[str, Sequence[str], None] = 'bbbcf69ab6be'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-	"""Upgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.create_table(
-		'categories',
-		sa.Column('id', sa.Integer(), nullable=False),
-		sa.Column('name', sa.String(length=100), nullable=False),
-		sa.Column('description', sa.Text(), nullable=True),
-		sa.Column(
-			'created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True
-		),
-		sa.Column(
-			'updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True
-		),
-		sa.PrimaryKeyConstraint('id'),
-		sa.UniqueConstraint('name'),
-	)
-	op.create_index(op.f('ix_categories_id'), 'categories', ['id'], unique=False)
-	op.create_table(
-		'questions',
-		sa.Column('id', sa.Integer(), nullable=False),
-		sa.Column('text', sa.String(length=255), nullable=False),
-		sa.Column('category_id', sa.Integer(), nullable=False),
-		sa.Column(
-			'created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True
-		),
-		sa.Column(
-			'updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True
-		),
-		sa.ForeignKeyConstraint(['category_id'], ['categories.id'], ondelete='CASCADE'),
-		sa.PrimaryKeyConstraint('id'),
-	)
-	op.create_index(op.f('ix_questions_id'), 'questions', ['id'], unique=False)
-	op.create_table(
-		'answers',
-		sa.Column('id', sa.Integer(), nullable=False),
-		sa.Column('text', sa.Text(), nullable=False),
-		sa.Column('is_correct', sa.Integer(), nullable=True),
-		sa.Column('question_id', sa.Integer(), nullable=False),
-		sa.Column(
-			'created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True
-		),
-		sa.Column(
-			'updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=True
-		),
-		sa.ForeignKeyConstraint(['question_id'], ['questions.id'], ondelete='CASCADE'),
-		sa.PrimaryKeyConstraint('id'),
-	)
-	op.create_index(op.f('ix_answers_id'), 'answers', ['id'], unique=False)
-	# ### end Alembic commands ###
-
-
-def downgrade() -> None:
-	"""Downgrade schema."""
-	# ### commands auto generated by Alembic - please adjust! ###
-	op.drop_index(op.f('ix_answers_id'), table_name='answers')
-	op.drop_table('answers')
-	op.drop_index(op.f('ix_questions_id'), table_name='questions')
-	op.drop_table('questions')
-	op.drop_index(op.f('ix_categories_id'), table_name='categories')
-	op.drop_table('categories')
-	# ### end Alembic commands ###
