+++ alembic/versions/f01182a0b275_create_initial_models.py
@@ -1,0 +1,182 @@
+"""create initial models
+
+Revision ID: f01182a0b275
+Revises:
+Create Date: 2026-06-22 11:20:24.420874
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = 'f01182a0b275'
+down_revision: Union[str, Sequence[str], None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+	"""Upgrade schema."""
+	# ### commands auto generated by Alembic - please adjust! ###
+	op.create_table(
+		'categories',
+		sa.Column('id', sa.Integer(), nullable=False),
+		sa.Column('name', sa.String(length=100), nullable=False),
+		sa.Column('description', sa.Text(), nullable=True),
+		sa.Column(
+			'created_at',
+			sa.DateTime(timezone=True),
+			server_default=sa.text('now()'),
+			nullable=False,
+		),
+		sa.Column(
+			'updated_at',
+			sa.DateTime(timezone=True),
+			server_default=sa.text('now()'),
+			nullable=False,
+		),
+		sa.PrimaryKeyConstraint('id'),
+		sa.UniqueConstraint('name'),
+	)
+	op.create_index(op.f('ix_categories_id'), 'categories', ['id'], unique=False)
+	op.create_table(
+		'users',
+		sa.Column('id', sa.Integer(), nullable=False),
+		sa.Column('email', sa.String(length=255), nullable=False),
+		sa.Column('phone_number', sa.String(length=16), nullable=True),
+		sa.Column('username', sa.String(length=100), nullable=False),
+		sa.Column('full_name', sa.String(length=255), nullable=True),
+		sa.Column('is_active', sa.Boolean(), nullable=False),
+		sa.Column(
+			'created_at',
+			sa.DateTime(timezone=True),
+			server_default=sa.text('now()'),
+			nullable=False,
+		),
+		sa.Column(
+			'updated_at',
+			sa.DateTime(timezone=True),
+			server_default=sa.text('now()'),
+			nullable=False,
+		),
+		sa.PrimaryKeyConstraint('id'),
+	)
+	op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
+	op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
+	op.create_index(op.f('ix_users_phone_number'), 'users', ['phone_number'], unique=True)
+	op.create_index(op.f('ix_users_username'), 'users', ['username'], unique=True)
+	op.create_table(
+		'games',
+		sa.Column('id', sa.Integer(), nullable=False),
+		sa.Column('name', sa.String(length=100), nullable=False),
+		sa.Column('min_players', sa.Integer(), nullable=False),
+		sa.Column('max_players', sa.Integer(), nullable=False),
+		sa.Column('max_timeout', sa.Integer(), nullable=False),
+		sa.Column('category_id', sa.Integer(), nullable=False),
+		sa.Column(
+			'created_at',
+			sa.DateTime(timezone=True),
+			server_default=sa.text('now()'),
+			nullable=False,
+		),
+		sa.Column(
+			'updated_at',
+			sa.DateTime(timezone=True),
+			server_default=sa.text('now()'),
+			nullable=False,
+		),
+		sa.Column('started_at', sa.DateTime(timezone=True), nullable=True),
+		sa.Column('ended_at', sa.DateTime(timezone=True), nullable=True),
+		sa.Column('creator_id', sa.Integer(), nullable=False),
+		sa.Column('private', sa.Boolean(), nullable=False),
+		sa.Column('password', sa.String(length=100), nullable=True),
+		sa.Column('winner', sa.Integer(), nullable=True),
+		sa.Column(
+			'state',
+			sa.Enum(
+				'WAITING', 'ACTIVE', 'FINISHED', name='gamestate', native_enum=False, length=50
+			),
+			nullable=False,
+		),
+		sa.ForeignKeyConstraint(
+			['category_id'],
+			['categories.id'],
+		),
+		sa.ForeignKeyConstraint(
+			['creator_id'],
+			['users.id'],
+		),
+		sa.ForeignKeyConstraint(
+			['winner'],
+			['users.id'],
+		),
+		sa.PrimaryKeyConstraint('id'),
+	)
+	op.create_index(op.f('ix_games_id'), 'games', ['id'], unique=False)
+	op.create_table(
+		'questions',
+		sa.Column('id', sa.Integer(), nullable=False),
+		sa.Column('text', sa.String(length=255), nullable=False),
+		sa.Column('category_id', sa.Integer(), nullable=False),
+		sa.Column('answers', sa.JSON(), nullable=False),
+		sa.Column('correct_answer', sa.Integer(), nullable=False),
+		sa.Column('approved', sa.Boolean(), nullable=False),
+		sa.Column(
+			'created_at',
+			sa.DateTime(timezone=True),
+			server_default=sa.text('now()'),
+			nullable=False,
+		),
+		sa.Column(
+			'updated_at',
+			sa.DateTime(timezone=True),
+			server_default=sa.text('now()'),
+			nullable=False,
+		),
+		sa.ForeignKeyConstraint(
+			['category_id'],
+			['categories.id'],
+		),
+		sa.PrimaryKeyConstraint('id'),
+	)
+	op.create_index(op.f('ix_questions_id'), 'questions', ['id'], unique=False)
+	op.create_table(
+		'game_players',
+		sa.Column('game_id', sa.Integer(), nullable=False),
+		sa.Column('user_id', sa.Integer(), nullable=False),
+		sa.ForeignKeyConstraint(['game_id'], ['games.id'], ondelete='CASCADE'),
+		sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
+		sa.PrimaryKeyConstraint('game_id', 'user_id'),
+	)
+	op.create_table(
+		'game_questions',
+		sa.Column('game_id', sa.Integer(), nullable=False),
+		sa.Column('question_id', sa.Integer(), nullable=False),
+		sa.ForeignKeyConstraint(['game_id'], ['games.id'], ondelete='CASCADE'),
+		sa.ForeignKeyConstraint(['question_id'], ['questions.id'], ondelete='CASCADE'),
+		sa.PrimaryKeyConstraint('game_id', 'question_id'),
+	)
+	# ### end Alembic commands ###
+
+
+def downgrade() -> None:
+	"""Downgrade schema."""
+	# ### commands auto generated by Alembic - please adjust! ###
+	op.drop_table('game_questions')
+	op.drop_table('game_players')
+	op.drop_index(op.f('ix_questions_id'), table_name='questions')
+	op.drop_table('questions')
+	op.drop_index(op.f('ix_games_id'), table_name='games')
+	op.drop_table('games')
+	op.drop_index(op.f('ix_users_username'), table_name='users')
+	op.drop_index(op.f('ix_users_phone_number'), table_name='users')
+	op.drop_index(op.f('ix_users_id'), table_name='users')
+	op.drop_index(op.f('ix_users_email'), table_name='users')
+	op.drop_table('users')
+	op.drop_index(op.f('ix_categories_id'), table_name='categories')
+	op.drop_table('categories')
+	# ### end Alembic commands ###
