+++ docker-compose.dev.yml
@@ -1,5 +1,3 @@
-version: '3.3'
-
 services:
   # FastAPI Backend
   backend:
